break down the Perception module further Game state management (saving/loading/creation) Rendering logic Movement and maze interaction Core game mechanics (walls, positions, etc.)
