+++ src/modules/perception/mod.rs
@@ -66,262 +66,7 @@ impl Perception {
         game.start()?;
         Ok(game)
     }
-    fn save_state(&self) -> Result<(), JsValue> {
-        let window = web_sys::window().expect("no global window exists");
-        let storage = window.local_storage()?.expect("no local storage exists");
-
-        // Save game state
-        let state = serde_wasm_bindgen::to_value(&self)?;
-        let state_json = js_sys::JSON::stringify(&state)?.as_string().unwrap();
-        storage.set_item("maze_state", &state_json)?;
-
-        // Save current time
-        storage.set_item("maze_time", &js_sys::Date::now().to_string())?;
-
-        // Save level separately
-        storage.set_item("maze_level", &self.level.to_string())?;
-
-        Ok(())
-    }
-    fn create_maze(size: usize, document: Document) -> Self {
-        let mut walls = vec![false; size * size * 4]; // Start with no walls
-                                                      // Add random walls
-        for i in 0..walls.len() {
-            walls[i] = Math::random() < 0.5;
-        }
-
-        let (waypoint1, key_position, waypoint2, door_position) = if size == 2 {
-            let key_pos = loop {
-                let pos = (
-                    (Math::random() * 2.0).floor() as usize,
-                    (Math::random() * 2.0).floor() as usize,
-                );
-                if pos != (0, 0) {
-                    break pos;
-                };
-            };
-            // Get random non-start, non-key position for door
-            let door_pos = loop {
-                let pos = (
-                    (Math::random() * 2.0).floor() as usize,
-                    (Math::random() * 2.0).floor() as usize,
-                );
-                if pos != (0, 0) && pos != key_pos {
-                    break pos;
-                }
-            };
-            (key_pos, key_pos, door_pos, door_pos)
-        } else {
-            let (mut key_pos, mut door_pos);
-            loop {
-                key_pos = (
-                    ((size as f64) * 0.6 + Math::random() * (size as f64) * 0.3).floor() as usize,
-                    ((size as f64) * 0.6 + Math::random() * (size as f64) * 0.3).floor() as usize,
-                );
-
-                door_pos = (
-                    ((size as f64) * 0.7 + Math::random() * (size as f64) * 0.2).floor() as usize,
-                    ((size as f64) * 0.7 + Math::random() * (size as f64) * 0.2).floor() as usize,
-                );
-
-                if key_pos != door_pos && key_pos != (0, 0) && door_pos != (0, 0) {
-                    break;
-                }
-            }
-            (
-                (
-                    (Math::random() * (size as f64)).floor() as usize,
-                    (Math::random() * (size as f64)).floor() as usize,
-                ),
-                key_pos,
-                (
-                    (Math::random() * (size as f64)).floor() as usize,
-                    (Math::random() * (size as f64)).floor() as usize,
-                ),
-                door_pos,
-            )
-        };
-
-        let mut game = Self {
-            size,
-            walls,
-            current_position: (0, 0),
-            key_position,
-            door_position,
-            visited: HashSet::new(),
-            has_key: false,
-            level: 1,
-            mazes_completed: 0,
-            document,
-            time_remaining: 300,
-            last_tick: js_sys::Date::now() / 1000.0,
-        };
-
-        // Clear path function - ensures a 2-cell wide path
-        fn clear_path(
-            walls: &mut Vec<bool>,
-            from: (usize, usize),
-            to: (usize, usize),
-            size: usize,
-        ) {
-            let mut current = from;
-            // Calculate minimum required path length (Manhattan distance * 1.5)
-            while current != to {
-                let dx = (to.0 as i32 - current.0 as i32).signum();
-                let dy = (to.1 as i32 - current.1 as i32).signum();
-                // Clear both current cell's wall and neighbor's wall
-                if dx != 0 {
-                    let wall_idx = (current.1 * size + current.0) * 4 + if dx > 0 { 1 } else { 3 };
-                    walls[wall_idx] = false;
-                    // Clear adjacent cell's opposite wall if not at edge
-                    if (dx > 0 && current.0 + 1 < size) || (dx < 0 && current.0 > 0) {
-                        let next_x = (current.0 as i32 + dx) as usize;
-                        let adj_wall_idx =
-                            (current.1 * size + next_x) * 4 + if dx > 0 { 3 } else { 1 };
-                        walls[adj_wall_idx] = false;
-
-                        // Always clear an escape route (up or down)
-                        let escape_dir = if current.1 > 0 { 0 } else { 2 }; // up if not at top, down otherwise
-                        walls[(current.1 * size + current.0) * 4 + escape_dir] = false;
-                        if escape_dir == 0 && current.1 > 0 {
-                            // Clear the corresponding wall in the cell above
-                            walls[((current.1 - 1) * size + current.0) * 4 + 2] = false;
-                        } else if escape_dir == 2 && current.1 + 1 < size {
-                            // Clear the corresponding wall in the cell below
-                            walls[((current.1 + 1) * size + current.0) * 4 + 0] = false;
-                        }
-                    }
-                    current.0 = (current.0 as i32 + dx) as usize;
-                } else if dy != 0 {
-                    let wall_idx = (current.1 * size + current.0) * 4 + if dy > 0 { 2 } else { 0 };
-                    walls[wall_idx] = false;
-                    // Clear adjacent cell's opposite wall if not at edge
-                    if (dy > 0 && current.1 + 1 < size) || (dy < 0 && current.1 > 0) {
-                        let next_y = (current.1 as i32 + dy) as usize;
-                        let adj_wall_idx =
-                            (next_y * size + current.0) * 4 + if dy > 0 { 0 } else { 2 };
-                        walls[adj_wall_idx] = false;
-                    }
-                    current.1 = (current.1 as i32 + dy) as usize;
-                }
-                // Ensure escape route from the destination
-                let escape_dirs = [(0, -1), (0, 1), (-1, 0), (1, 0)]; // up, down, left, right
-                for (dx, dy) in escape_dirs.iter() {
-                    let next_x = to.0 as i32 + dx;
-                    let next_y = to.1 as i32 + dy;
-                    if next_x >= 0 && next_x < size as i32 && next_y >= 0 && next_y < size as i32 {
-                        let wall_idx = (to.1 * size + to.0) * 4
-                            + if *dy < 0 {
-                                0
-                            } else if *dx > 0 {
-                                1
-                            } else if *dy > 0 {
-                                2
-                            } else {
-                                3
-                            };
-                        walls[wall_idx] = false;
-                    }
-                }
-            }
-        }
-        // Create path through waypoints
-        clear_path(&mut game.walls, (0, 0), waypoint1, size);
-        clear_path(&mut game.walls, waypoint1, key_position, size);
-        clear_path(&mut game.walls, key_position, waypoint2, size);
-        clear_path(&mut game.walls, waypoint2, door_position, size);
-        game.visited.insert((0, 0));
-        game
-    }
-    pub fn render(&self) -> Result<(), JsValue> {
-        let maze = self.document.get_element_by_id("maze").unwrap();
-    
-        // Only regenerate grid if size changed
-        if maze.children().length() as usize != self.size * self.size {
-            maze.set_attribute(
-                "style",
-                &format!("grid-template-columns: repeat({}, 60px)", self.size),
-            )?;
-            
-            // Clear existing content safely
-            while let Some(child) = maze.first_child() {
-                maze.remove_child(&child)?;
-            }
-    
-            // Create cells only once
-            for _ in 0..(self.size * self.size) {
-                let cell = self.document.create_element("div")?;
-                cell.set_class_name("cell");
-                let span = self.document.create_element("span")?;
-                let content = self.document.create_text_node("");
-                cell.append_child(&content)?;
-                cell.append_child(&span)?;
-                maze.append_child(&cell)?;
-            }
-        }
-    
-        // Update existing cells
-        for y in 0..self.size {
-            for x in 0..self.size {
-                let index = (y * self.size + x) as u32;
-                if let Some(cell) = maze.children().item(index) {
-                    self.update_cell_state(&cell, x, y)?;
-                }
-            }
-        }
-    
-        // Update stats
-        if let Some(level_el) = self.document.get_element_by_id("level") {
-            level_el.set_text_content(Some(&self.level.to_string()));
-        }
-        if let Some(completed_el) = self.document.get_element_by_id("completed") {
-            completed_el.set_text_content(Some(&self.mazes_completed.to_string()));
-        }
-        if let Some(timer_el) = self.document.get_element_by_id("timer") {
-            let minutes = self.time_remaining / 60;
-            let seconds = self.time_remaining % 60;
-            timer_el.set_text_content(Some(&format!("{}:{:02}", minutes, seconds)));
-        }
-        Ok(())
-    }
-    
-    fn update_cell_state(&self, cell: &Element, x: usize, y: usize) -> Result<(), JsValue> {
-        // Reset base class
-        cell.set_class_name("cell");
-        
-        // Update state classes
-        if self.visited.contains(&(x, y)) {
-            cell.class_list().add_1("visited")?;
-        }
-        if (x, y) == self.current_position {
-            cell.class_list().add_1("current")?;
-            // Ensure span exists for pseudo-elements
-            if cell.children().length() == 0 {
-                let span = self.document.create_element("span")?;
-                cell.append_child(&span)?;
-            }
-        }
-    
-        // Update content
-        let content = if (x, y) == self.key_position && !self.has_key {
-            "🔑"
-        } else if (x, y) == self.current_position && self.has_key {
-            "🔑"
-        } else if (x, y) == self.door_position {
-            "🚪"
-        } else {
-            ""
-        };
-    
-        // Update text content if it's different
-        if let Some(first_child) = cell.first_child() {
-            if first_child.text_content().unwrap_or_default() != content {
-                first_child.set_text_content(Some(content));
-            }
-        }
-    
-        Ok(())
-    }    #[wasm_bindgen]
+    #[wasm_bindgen]
     pub fn start(&mut self) -> Result<(), JsValue> {
         let game_state = Rc::new(RefCell::new(self.clone()));
 
@@ -424,99 +169,6 @@ impl Perception {
         f.forget();
         console::log_1(&"Setup complete".into());
         Ok(())
-    }
-    fn is_adjacent(&self, x: usize, y: usize) -> bool {
-        let current_x = self.current_position.0;
-        let current_y = self.current_position.1;
-
-        // Check if target position is adjacent (up, down, left, right)
-        let dx = if x >= current_x {
-            x - current_x
-        } else {
-            current_x - x
-        };
-        let dy = if y >= current_y {
-            y - current_y
-        } else {
-            current_y - y
-        };
-
-        // Only one coordinate can change by 1, the other must be 0
-        (dx == 1 && dy == 0) || (dx == 0 && dy == 1)
-    }
-
-    fn get_wall_index(&self, from_x: usize, from_y: usize, to_x: usize, to_y: usize) -> usize {
-        let cell_walls = 4; // each cell has 4 possible walls
-        let base_index = (from_y * self.size + from_x) * cell_walls;
-
-        if to_x > from_x {
-            base_index + 1 // right wall
-        } else if to_x < from_x {
-            base_index + 3 // left wall
-        } else if to_y > from_y {
-            base_index + 2 // bottom wall
-        } else {
-            base_index + 0 // top wall
-        }
-    }
-    #[wasm_bindgen]
-    pub fn try_move(&mut self, x: usize, y: usize) -> i32 {
-        let result = self.try_move_internal(x, y);
-        if result != 0 {
-            self.save_state().unwrap_or_else(|_| {
-                console::log_1(&"Failed to save game state".into());
-            });
-        }
-        result
-    }
-    fn try_move_internal(&mut self, x: usize, y: usize) -> i32 {
-        if !self.is_adjacent(x, y) {
-            return 0;
-        }
-
-        let wall_idx = self.get_wall_index(self.current_position.0, self.current_position.1, x, y);
-
-        // Block access to door position if key not collected
-        if (x, y) == self.door_position && !self.has_key {
-            return 0;
-        }
-
-        if self.walls[wall_idx] {
-            self.reset_position();
-            return -1;
-        }
-
-        self.current_position = (x, y);
-        self.visited.insert((x, y));
-
-        if (x, y) == self.key_position {
-            self.has_key = true;
-            // When key is collected, make door accessible
-            let door_x = self.door_position.0;
-            let door_y = self.door_position.1;
-            let base_idx = (door_y * self.size + door_x) * 4;
-            for i in 0..4 {
-                self.walls[base_idx + i] = false;
-            }
-        }
-
-        if (x, y) == self.door_position && self.has_key {
-            // Simplified level up - increase size immediately
-            self.size += 1;
-            self.level += 1;
-            let new_game = Self::create_maze(self.size, self.document.clone());
-            self.walls = new_game.walls;
-            self.current_position = (0, 0);
-            self.key_position = new_game.key_position;
-            self.door_position = new_game.door_position;
-            self.visited.clear();
-            self.visited.insert((0, 0));
-            self.has_key = false;
-            self.time_remaining = 300;
-            self.last_tick = js_sys::Date::now() / 1000.0;
-            return 2;
-        }
-        1
     }
     #[wasm_bindgen]
     pub fn reset(&mut self) {
